Module側でスコープを指定するように変更
