+++ app/src/main/java/com/example/gourmetsearcher/di/keywordhistory/KeyWordHistoryModule.kt
@@ -5,6 +5,8 @@ import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
 import com.example.gourmetsearcher.manager.PreferencesManager
+import com.example.gourmetsearcher.repository.KeyWordHistoryRepository
+import com.example.gourmetsearcher.repository.KeyWordHistoryRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -12,12 +14,13 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
-/** プリファレンスのモジュール */
+/** 検索履歴のモジュール */
 @Module
 @InstallIn(SingletonComponent::class)
-object PreferencesModule {
+object KeyWordHistoryModule {
     /** データストアを提供する */
-    private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "HistoryPrefs")
+    private val Context.dataStore: DataStore<Preferences>
+    by preferencesDataStore(name = "HistoryPrefs")
 
     /**
      * データストアを提供する
@@ -25,6 +28,7 @@ object PreferencesModule {
      * @return データストア
      */
     @Provides
+    @Singleton
     fun provideDataStore(
         @ApplicationContext context: Context,
     ): DataStore<Preferences> = context.dataStore
@@ -36,5 +40,16 @@ object PreferencesModule {
      */
     @Provides
     @Singleton
-    fun providePreferencesManger(dataStore: DataStore<Preferences>): PreferencesManager = PreferencesManager(dataStore)
+    fun providePreferencesManger(dataStore: DataStore<Preferences>): PreferencesManager
+    = PreferencesManager(dataStore)
+
+    /**
+     * 検索履歴のリポジトリを提供する
+     * @param preferencesManager プリファレンスマネージャ
+     * @return 検索履歴のリポジトリ
+     */
+    @Provides
+    @Singleton
+    fun provideKeyWordHistory(preferencesManager: PreferencesManager): KeyWordHistoryRepository =
+        KeyWordHistoryRepositoryImpl(preferencesManager)
 }
