+++ app/src/main/java/com/example/gourmetsearcher/di/network/NetWorkModule.kt
@@ -2,6 +2,10 @@ package com.example.gourmetsearcher.di.network
 
 import android.content.Context
 import com.example.gourmetsearcher.R
+import com.example.gourmetsearcher.manager.CacheManager
+import com.example.gourmetsearcher.repository.RestaurantRepository
+import com.example.gourmetsearcher.repository.RestaurantRepositoryImpl
+import com.example.gourmetsearcher.service.HotPepperGourmetApiService
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
@@ -11,6 +15,7 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
+import javax.inject.Singleton
 
 /** NetWorkのモジュール */
 @Module
@@ -21,6 +26,7 @@ object NetWorkModule {
      * @return Moshi
      */
     @Provides
+    @Singleton
     fun provideMoshi(): Moshi =
         Moshi
             .Builder()
@@ -34,6 +40,7 @@ object NetWorkModule {
      * @return Retrofit
      */
     @Provides
+    @Singleton
     fun provideRetrofit(
         @ApplicationContext context: Context,
         moshi: Moshi,
@@ -43,4 +50,28 @@ object NetWorkModule {
             .baseUrl(context.getString(R.string.restaurant_list_hot_pepper_url))
             .addConverterFactory(MoshiConverterFactory.create(moshi))
             .build()
+
+    /**
+     *  HotPepperGourmetApiServiceを提供
+     * @param retrofit Retrofit
+     * @return HotPepperGourmetApiService
+     */
+    @Provides
+    @Singleton
+    fun provideNetWorkServiceModule(
+        retrofit: Retrofit
+    ): HotPepperGourmetApiService = retrofit.create(HotPepperGourmetApiService::class.java)
+
+    /**
+     * RestaurantRepositoryを提供
+     * @param service   HotPepperGourmetApiService
+     * @param cacheManager CacheManager
+     * @return RestaurantRepository
+     */
+    @Provides
+    @Singleton
+    fun provideRestaurantRepository(
+        service: HotPepperGourmetApiService,
+        cacheManager: CacheManager,
+    ): RestaurantRepository = RestaurantRepositoryImpl(service, cacheManager)
 }
